+++ src/main/java/com/fiap/fast_food_tc/adapter/dto/product/ProductResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Builder
 @Data
 @AllArgsConstructor
@@ -14,7 +16,7 @@ public class ProductResponse {
     private Integer productId;
     private String name;
     private Integer quantity;
-    private Double productValue;
+    private BigDecimal productValue;
     private Boolean isAvailable;
     private Integer categoryId;
     private String description;
